+++ switch_my_loan/switch_my_loan/report/sales_manager_and_status_wise_leads/sales_manager_and_status_wise_leads.js
@@ -1,0 +1,16 @@
+frappe.query_reports["Sales Manager and status wise Leads"] = {
+    "filters": [
+        {
+            "fieldname":"from_date",
+            "label":("From Date"),
+            "fieldtype":"Date",
+            "default":frappe.datetime.month_start()
+        },
+        {
+            "fieldname":"to_date",
+            "label":("To Date"),
+            "fieldtype":"Date",
+            "default":frappe.datetime.month_end()
+        },
+]
+}